Update to Openwall 1.1.3 C source
This adds support for the $2b$ prefix which will be used in future
versions of OpenBSD. See

http://marc.info/?l=openbsd-misc&m=139320023202696&q=raw

or

http://undeadly.org/cgi?action=article&sid=20140224132743

for the background on why it was introduced.

This commit

* Adds the updated C source.
* Updates the project description to clarify the version used and which
  modifications have been made to the openwall source, and provides a
  link to the original source website.
* Raises the upper bound on the 'entropy' dependency to avoid cabal
  problems.

In future, the default algorithm should probably be switched to "$2b$
since this will allow for maximum compatibility between systems, but it’s
probably best to wait until other libraries also have support.